+++ resources/views/components/toolbar/sorting.blade.php
@@ -16,17 +16,17 @@ UI components used:
       </x-lv-select-button>
     </x-slot>
 
-    <x-lv-drop-down.header label="Sort by" />
+    <x-lv-drop-down.header label="{{ __('Sort by') }}" />
 
     {{-- Each sortable item --}}
     @foreach ($sortableBy as $title => $column)
       <a href="#!"
         wire:click.prevent="sort('{{ $column }}')"
         title="{{ __('Sort by') }} {{ $title }} {{ __($sortOrder == 'asc' ? 'descending' : 'ascending') }}"
-        class="group flex items-center px-4 py-2 hover:bg-gray-100 hover:text-gray-900 text-sm gap-3">
+        class="group flex items-center px-4 py-2 hover:bg-gray-100 hover:text-gray-900 text-sm gap-3"
+      >
         @if ($sortBy === $column)
-          <i data-feather="arrow-{{ $sortOrder === 'asc' ? 'up' : 'down' }}"
-            class="text-gray-900 h-4 w-4"></i>
+          <i data-feather="arrow-{{ $sortOrder === 'asc' ? 'up' : 'down' }}" class="text-gray-900 h-4 w-4"></i>
         @else
           <i class="text-gray-900 h-4 w-4"></i>
         @endif