+++ resources/views/view/layout.blade.php
@@ -3,4 +3,7 @@
   <x-lv-alerts-handler />
 
   {{ $slot }}
+
+  {{-- Confirmation message alert --}}
+  <x-lv-confirmation-message />
 </div>