+++ src/LaravelViews.php
@@ -32,6 +32,9 @@ class LaravelViews
             'modal' => 'modal',
             'form.checkbox' => 'checkbox',
             'form.input' => 'input',
+            'tooltip' => 'tooltip',
+            'confirmation-message' => 'confirmation-message',
+            'loading' => 'loading'
         ];
     }