+++ src/LaravelViewsServiceProvider.php
@@ -14,12 +14,13 @@ use LaravelViews\UI\Variants;
 use LaravelViews\UI\Header;
 use Illuminate\Support\ServiceProvider;
 use Illuminate\Support\Facades\Blade;
-use Illuminate\Support\Str;
 use LaravelViews\Console\GridViewMakeCommand;
 use LaravelViews\Console\ListViewMakeCommand;
 use LaravelViews\Console\MakeViewCommand;
 use LaravelViews\Data\Contracts\Sortable;
 use LaravelViews\Data\TableViewSortData;
+use LaravelViews\Macros\LaravelViewsTestMacros;
+use LaravelViews\Macros\StrMacros;
 use LaravelViews\Views\Components\DynamicComponent;
 
 class LaravelViewsServiceProvider extends ServiceProvider
@@ -164,16 +165,15 @@ class LaravelViewsServiceProvider extends ServiceProvider
 
     private function macros()
     {
-        Str::macro('classNameAsSentence', function ($className) {
-            $intermediate = preg_replace('/(?!^)([[:upper:]][[:lower:]]+)/', ' $0', $className);
-            $titleStr = preg_replace('/(?!^)([[:lower:]])([[:upper:]])/', '$1 $2', $intermediate);
+        $macros = [
+            LaravelViewsTestMacros::class,
+            StrMacros::class
+        ];
 
-            return $titleStr;
-        });
-
-        Str::macro('camelToDash', function ($str) {
-            return strtolower(preg_replace('%([a-z])([A-Z])%', '\1-\2', $str));
-        });
+        foreach ($macros as $macroClass) {
+            $macro = new $macroClass;
+            $macro->register();
+        }
 
         return $this;
     }