+++ src/UI/UI.php
@@ -6,6 +6,14 @@ use Illuminate\Support\Facades\View;
 
 class UI
 {
+    public function editable($model, $field)
+    {
+        return $this->component('laravel-views::components.editable', [
+            'model' => $model,
+            'field' => $field
+        ]);
+    }
+
     public function badge($title, $type = 'default')
     {
         return $this->component('laravel-views::components.badge', [